// src/app/types.ts
export type TApplicationMetadata = {
  name: string
  website: string
  description: string
  email: string
  base64Image: string
}

export type TNewApplication = TApplicationMetadata & {
  requestedAmount: bigint
  recipientAddress: `0x${string}`
  profileId?: `0x${string}`
  profileName?: string
}